+++ src/main/java/by/ganevich/io/commands/DeleteClientCommand.java
@@ -1,16 +1,16 @@
 package by.ganevich.io.commands;
 
+import by.ganevich.dto.ClientDto;
 import by.ganevich.entity.Client;
 import by.ganevich.io.CommandDescriptor;
 import by.ganevich.io.CommandResult;
+import by.ganevich.mapper.interfaces.ClientMapper;
 import by.ganevich.service.ClientService;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 
-import javax.validation.constraints.NotEmpty;
-import javax.validation.constraints.Pattern;
-import javax.validation.constraints.Size;
+import javax.validation.Valid;
 import java.util.Map;
 
 @Component
@@ -20,11 +20,10 @@ public class DeleteClientCommand extends BaseCommand {
 
     private final String commandName = "deleteClient";
     private final ClientService clientService;
+    private final ClientMapper clientMapper;
 
-    @Pattern(regexp = "[A-Z][a-z]*", message = "Client name must start with a capital letter")
-    @Size(min = 2, max = 25, message = "Name length must be between 2 and 25")
-    @NotEmpty(message = "Name must not be empty")
-    private String clientName;
+    @Valid
+    private ClientDto clientDto;
 
 
     @Override
@@ -35,19 +34,27 @@ public class DeleteClientCommand extends BaseCommand {
 
     @Override
     public CommandResult doExecute(Map<String, String> parameters) {
-        Client client = clientService.findClientByName(parameters.get("clientName"));
 
-        clientService.removeClient(client);
+        String result;
+
+        Client client = clientMapper.toEntity(clientDto);
+        if (client == null) {
+            result = "Client already removed!";
+        } else {
+            clientService.removeClient(client);
+            result = "Client is successfully removed!";
+        }
 
         CommandResult commandResult = new CommandResult();
-        String result = "Client removed!";
         commandResult.setResult(result);
         return commandResult;
     }
 
     @Override
-    public ICommand setParameters(CommandDescriptor commandDescriptor) {
-        this.clientName = commandDescriptor.getParameters().get("clientName");
+    public ICommand setDto(CommandDescriptor commandDescriptor) {
+        ClientDto clientDto = clientMapper
+                .toDto(clientService.findClientByName(commandDescriptor.getParameters().get("clientName")));
+        this.clientDto = clientDto;
         return this;
     }
 }
